HKDF: Add explicit cast when combining size_t & int
Add an explicit cast when combining size_t and int as in the case of
assignments or comparisons to the struct hkdf_ctx mdlen member.

Tested:
- Arch Linux
- CentOS Stream 8, 8.3.2011, 7.9.2009, 6.10
- Debian sid, 10.8, 9.13, 8.11, 7.11, 6.0.10, 5.0.10, 4.0
- Fedora 33, 32, 31
- FreeBSD 12.2, 11.4
- NetBSD 9.1, 9.0, 8.1
- OpenBSD 6.8, 6.7, 6.6
- openSUSE 15.2, 15.1
- Raspbian 10.8 [armv7l]
- Ubuntu 20.10, 20.04.2 LTS, 18.04.5 LTS, 16.04.7 LTS, 14.04.6 LTS, 12.04.5 LTS

+++ src/common/hkdf.c
@@ -244,8 +244,8 @@ hkdf (hkdf_ctx_t *ctxp, void *dst, size_t *dstlenp)
     }
     /*  Compute length of hash function output.
      */
-    ctxp->mdlen = mac_size (ctxp->md);
-    if (ctxp->mdlen == -1) {
+    ctxp->mdlen = (size_t) mac_size (ctxp->md);
+    if (ctxp->mdlen == (size_t) -1) {
         errno = EINVAL;
         return -1;
     }
